Set up test training script.
