+++ .gitignore
@@ -138,3 +138,6 @@ dmypy.json
 src/optim/sphinx
 data/
 src/training_output
+
+# Tensorboard
+runs/
